requirements.txt
lxml>=4.9.1
latex2mathml>=3.75
PySide6-Essentials==6.4.0.1